Removes un-necessary test class
